Update README.md
Designed as a basic back-end software for use in a 
store like environment all we had to do was make 
a system that could add remove, edit items and 
have a shopping cart, a bunch of other features. 
I added a .txt file to store the changes. I've also added 
some things like if you checkout before you quit it'll 
just say Thank you but if you quit before checking out 
it shows the checkout.
The main file is ims.py